implement /transactions/:id/ GET and PUT for edits
We had to embed the Entry model into Transaction. It was causing
too many problems with no benefit.

The Transaction editing form was moved out to a partial so that
it could be reused both in the index and the :id/edit views.
The forms and index should now handle unlimited entries.